requirements.txt
feather-format==0.4.0
mock==2.0.0
numpy==1.16.1
pandas==0.24.1
pbr==5.1.2
psycopg2-binary==2.7.7
pyarrow==0.12.1
python-dateutil==2.8.0
pytz==2018.9
six==1.12.0
SQLAlchemy==1.2.18
wrds==3.0.6
